+++ CI/SCRIPTS/test_broken_s2.py
@@ -16,7 +16,7 @@ def test_broken_s2():
     """Function testing the support of broken Sentinel-2 sensor"""
     res = 10.0 * 100
 
-    # Broken MTD
+    # # Broken MTD
     broken_mtd = broken_s2_path().joinpath(
         "S2A_MSIL2A_20170331T103021_N0001_R108_T32UMV_20190508T053047.SAFE"
     )
@@ -33,3 +33,24 @@ def test_broken_s2():
     # Invalid tests
     with pytest.raises(InvalidProductError):
         broken_mtd_prod.read_mtd()
+
+    # Broken DETFOO
+    broken_detfoo = broken_s2_path().joinpath(
+        "S2A_MSIL2A_20170331T103021_N0001_R108_T32ULU_20190508T053054.SAFE"
+    )
+
+    broken_detfoo_prod = READER.open(broken_detfoo)
+
+    assert broken_detfoo_prod is not None
+    LOGGER.info(broken_detfoo_prod)
+    LOGGER.info(broken_detfoo_prod.bands)
+
+    broken_detfoo_prod.load(
+        RED, resolution=res, clean_optical="clean"
+    )  # Not corrupted band
+
+    # Invalid tests
+    with pytest.raises(InvalidProductError):
+        broken_detfoo_prod.load(
+            NIR, resolution=res, clean_optical="nodata"
+        )  # Corrupted band